Renamed test buffer to testBuffer
