+++ sdk/attestation/attestation/package.json
@@ -11,7 +11,7 @@
     "./dist-esm/src/utils/textEncoding.js": "./dist-esm/src/utils/textEncoding.browser.js",
     "./dist-esm/src/utils/buffer.js": "./dist-esm/src/utils/buffer.browser.js",
     "./dist-esm/test/utils/base64url.js": "./dist-esm/test/utils/base64url.browser.js",
-    "./dist-esm/test/utils/buffer.js": "./dist-esm/test/utils/buffer.browser.js",
+    "./dist-esm/test/utils/safeBuffer.js": "./dist-esm/test/utils/safeBuffer.browser.js",
     "./dist-esm/test/utils/env.js": "./dist-esm/test/utils/env.browser.js",
     "./dist-esm/samples-dev/utils/base64url.js": "./dist-esm/samples-dev/utils/base64url.browser.js"
   },
