+++ sdk/attestation/attestation/test/utils/decodeJWT.ts
@@ -4,7 +4,7 @@
 // typed implementation of jwsDecode, copied from here: https://github.com/auth0/node-jws/blob/master/lib/verify-stream.js
 /* eslint-disable */
 
-import { Buffer } from "./Buffer";
+import { Buffer } from "./safeBuffer";
 
 export function decode(
   jwsSig: string,
