+++ sdk/attestation/attestation/test/utils/safeBuffer.browser.ts
