+++ sdk/attestation/attestation/test/utils/safeBuffer.ts
